requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pymongo==4.4.1
bcrypt==4.3.0
PyJWT==2.7.0
celery==5.3.1
redis==4.6.0
matplotlib==3.7.1
dashscope==1.13.6